requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.6

numpy==2.0.2
pandas==2.2.3
scipy==1.14.1

requests==2.32.3
pystac-client==0.8.3
planetary-computer==1.0.0

rasterio==1.3.10
pyproj==3.6.1
gunicorn

google-genai==1.14.0
